Use optional dummy OpenImageIO api
So that we can compile the code without having to link OpenImageIO.
I am not sure that this is the proper way, maybe it would be better to
simply remove the image input/output code and assume that we always use
the callbacks for that.

In order to prevent some compilation issues, I renamed the OIIO
namespace to IMG for the image functions: that way we can compile the
core without OpenImageIO but still use it for arguments parsing.