Don't use OpenImageIO for ustring and TypeDesc
Instead we base ustring on std::string, and use our own implementation
of TypeDesc.

That way we can remove the OIIO dependency from the core.

There might be some performances issues from using std::string instead
of OIIO::ustring.  I am not sure if that is a real problem.  If so we
could have our own implementation of it in cycles.